+++ src/PhoneNumbers.FluentValidation/CHANGELOG.md
@@ -1,2 +1,2 @@
 ﻿1.0.0
-  - Initial release with the support FluentValidation for the validation of EmailAddress value object.
+  - Initial release with the support FluentValidation for the validation of PhoneNumber value object.